ecommerce-bigdata-platform/backend/requirements.txt
fastapi==0.115.13
uvicorn==0.32.1
motor==3.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic==2.10.4
confluent-kafka==2.3.0
python-dotenv==1.0.1
cloudinary==1.44.1
pymongo>=4.9,<5.0
bcrypt==4.1.2
requests==2.31.0
